requirements.txt
langchain==0.3.1
langchain_community==0.3.1
langchain_huggingface==0.1.0
python-dotenv==1.0.1
streamlit==1.38.0
